helpers/audio: extract play, playAll
