+++ src/components/exercises/ExplanationText.js
@@ -2,9 +2,10 @@ import React from 'react';
 import { View, Text } from 'react-native';
 
 import speakerImage from '../../assets/images/speaker.png';
+import { play } from '../../helpers/audio';
 import { GREEN } from '../../styles/colors';
 import { DEFAULT } from '../../styles/text';
-import { loadSound, play } from '../helpers/audio';
+import { loadSound } from '../helpers/audio';
 import { RoundButton } from '../Components';
 
 const styles = {
