+++ src/components/exercises/HearWord.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import { View } from 'react-native';
 
 import speakerImage from '../../assets/images/speaker.png';
-
+import { play } from '../../helpers/audio';
 import { PRIMARY } from '../../styles/colors';
 import { RoundButton, RoundTextButton } from '../Components';
-import { loadSound, play } from '../helpers/audio';
+import { loadSound } from '../helpers/audio';
 
 class HearWord extends Component {
   constructor(props) {
