+++ src/components/exercises/MissingText.js
@@ -3,10 +3,10 @@ import React from 'react';
 import { View, Text } from 'react-native';
 
 import speakerImage from '../../assets/images/speaker.png';
-
+import { playAll } from '../../helpers/audio';
 import { PRIMARY } from '../../styles/colors';
 import { DEFAULT } from '../../styles/text';
-import { loadSounds, playAll } from '../helpers/audio';
+import { loadSounds } from '../helpers/audio';
 import { RoundImageWithButton, TextPicker } from '../Components';
 import Video from '../common/Video';
 const mapIndexed = addIndex(map);
