+++ src/components/helpers/audio.js
@@ -2,35 +2,6 @@ import { Audio } from 'expo';
 import { forEach, map } from 'ramda';
 import React, { Component } from 'react';
 
-export const play = sound =>
-  new Promise(resolve => {
-    sound.playAsync();
-
-    sound.setOnPlaybackStatusUpdate(status => {
-      if (status.didJustFinish) {
-        sound.stopAsync();
-        resolve();
-      }
-    });
-  });
-
-export const playAll = (sounds, delay = 1000) =>
-  new Promise(resolve => {
-    const playAllRecursive = ([sound, ...rest]) => {
-      play(sound).then(() => {
-        if (rest.length === 0) {
-          return resolve();
-        }
-
-        setTimeout(() => {
-          playAllRecursive(rest, delay);
-        }, delay);
-      });
-    };
-
-    playAllRecursive(sounds);
-  });
-
 export const createLoadSounds = Audio => C => {
   Audio.setIsEnabledAsync(true);
